+++ components/synth_core/synth_ui/ui_screen_notefx.c
@@ -47,7 +47,7 @@ uint8_t notefx_menu_item_count(void)
 
 const menu_item_view_t *notefx_menu_build_items(void)
 {
-    for (uint8_t i = 0; i < NFX_COUNT; i++) {
+    for (int i = 0; i < NFX_COUNT; i++) {
         s_nfx_items[i].value[0] = '\0';
     }
